Uploaded site project files v.1.0 (NO CONFIG)
Code unchanged will not function, like, AT ALL. This is due to the fact that I'm leaving out sensitive Firebase configuration data. It's all on the client-side (as it should be for Firebase/Firestore, I think?) but I haven't added any auth rules, so it's open to the public editing even on-site currently. I just didn't want to add the API keys etc. to plain text and make it EVEN MORE accessible to the public 💀